+++ examples/summarization/ipu_config.json
@@ -1,12 +1,12 @@
 {
-  "embedding_serialization_factor": 1,
+  "embedding_serialization_factor": 5,
   "recompute_checkpoint_every_layer": true,
   "optimizer_state_offchip": true,
   "replicated_tensor_sharding": true,
   "enable_half_first_order_momentum": true,
   "enable_half_partials": true,
 
-  "device_iterations": 2,
+  "device_iterations": 1,
   "inference_device_iterations": 5,
   "replication_factor": 4,
   "inference_replication_factor": 4,
@@ -15,5 +15,5 @@
 
   "ipus_per_replica": 4,
   "layers_per_ipu": [0, 4, 4, 4],
-  "matmul_proportion": [0.25, 0.2, 0.2, 0.25]
+  "matmul_proportion": [0.05, 0.2, 0.2, 0.25]
 }
