+++ optimum/graphcore/models/vit/modeling_vit.py
@@ -12,38 +12,14 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-import torch
-import torch.nn as nn
-
 import poptorch
 import transformers
 from optimum.utils import logging
 
-from ...modeling_utils import PipelineMixin, register
+from ...modeling_utils import PipelineMixin, _get_layer_ipu, recomputation_checkpoint, register
 
 
 logger = logging.get_logger(__name__)
-
-
-def _get_layer_ipu(layers_per_ipu):
-    # List of the IPU Id for each encoder layer
-    layer_ipu = []
-    for ipu, n_layers in enumerate(layers_per_ipu):
-        layer_ipu += [ipu] * n_layers
-    return layer_ipu
-
-
-def recomputation_checkpoint(module: nn.Module) -> torch.utils.hooks.RemovableHandle:
-    """Annotates the output of a module to be checkpointed instead of
-    recomputed"""
-
-    def recompute_outputs(module, inputs, outputs):
-        if type(outputs) is tuple:
-            return tuple(poptorch.recomputationCheckpoint(y) for y in outputs)
-        else:
-            return poptorch.recomputationCheckpoint(outputs)
-
-    return module.register_forward_hook(recompute_outputs)
 
 
 @register(transformers.ViTForImageClassification)
